requirements.txt
ccxt
sanpy
numpy
pandas
scipy
streamlit
plotly
